feat(client api): exposed more apis to python
added HcAgentRegisterIcon, HcAgentRegisterIconName, and HcAgentRegisterInitialize
to the python api to allow setting custom session icons for the graph
and added the HcAgentRegisterInitialize api to set a python function
callback to be called every time a new agent is initialized to the
server. A little of refactoring has been done as well.